create(#26): Article and ArticleComment
* feat(#8, #9): 게시글, 댓글 API 구현을 위한 디렉토리 생성

* feat(#25): 게시글 도메인 설계를 위한 Entity 생성

* feat(#25): H2, JPA, Log를 사용하기 위한 properties 설정

* remove(#25): 불필요한 데이터 삭제

* config(#25): build.gradle 업데이트

* config(#25): gitignore 설정

* feat(#25): Controller, Service 생성

* feat(#25): Account 엔티티 생성

* feat(#25): Article 엔티티 생성

* feat(#25): ArticleComment 엔티티 생성

* feat(#25): ArticleHashtag 엔티티 생성

* feat(#25): HashTag 엔티티 생성

* feat(#25): Role enum 생성

* feat(#25):Update .gitignore

* feat(#25): comment 추가

* feat(#25): 룰을 지움에 따라 TODO 삭제

* feat(#25): Update .gitignore && EOL 추가

* remove(#25): 불필요한 의존성 제거

* refactor(#25): 적합한 자료구조 사용으로 변환

* refactor(#25): Wrapper 클래스 사용

* refactor(#25): MappedSuperclass 적용

* refactor(#25): access 범위 지정

- PROTECTED
