+++ .gitignore
@@ -390,3 +390,4 @@ obj/
 !/gradle/wrapper/gradle-wrapper.jar
 
 # End of https://www.toptal.com/developers/gitignore/api/java,macos,windows,linux,vue,vuejs,visualstudiocode,android,androidstudio,gradle,intellij+all
+
