+++ server/.gitignore
@@ -36,5 +36,5 @@ out/
 ### VS Code ###
 .vscode/
 
-
 src/main/resources/application-dev.properties
+src/main/resources/application-real.properties
