+++ server/src/main/java/board/api/config/AppConfig.java
@@ -1,0 +1,21 @@
+package board.api.config;
+
+import board.api.modules.account.mapper.AccountMapper;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.factory.PasswordEncoderFactories;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class AppConfig {
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
+    }
+
+    @Bean
+    public AccountMapper memberMapper() {
+        return new AccountMapper();
+    }
+}
