+++ server/src/main/java/board/api/config/AppProperties.java
@@ -1,0 +1,15 @@
+package board.api.config;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.ConstructorBinding;
+
+@Getter
+@RequiredArgsConstructor
+@ConstructorBinding
+@ConfigurationProperties("app")
+public class AppProperties {
+
+    private final String host;
+}
