+++ server/src/main/java/board/api/config/CacheConfig.java
@@ -1,0 +1,25 @@
+package board.api.config;
+
+import java.util.Arrays;
+import org.springframework.cache.CacheManager;
+import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.cache.concurrent.ConcurrentMapCache;
+import org.springframework.cache.support.SimpleCacheManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+@EnableCaching
+@Configuration
+public class CacheConfig {
+
+    public static final String ACCOUNT_CACHE_NAME = "accountTemp";
+
+    @Bean
+    public CacheManager cacheManager() {
+        SimpleCacheManager cacheManager = new SimpleCacheManager();
+        cacheManager.setCaches(Arrays.asList(
+            new ConcurrentMapCache(ACCOUNT_CACHE_NAME)
+        ));
+        return cacheManager;
+    }
+}
