+++ server/src/main/java/board/api/exception/InvalidSignUpRequestException.java
@@ -1,0 +1,8 @@
+package board.api.exception;
+
+public class InvalidSignUpRequestException extends IllegalArgumentException {
+
+    public InvalidSignUpRequestException(String message) {
+        super(message);
+    }
+}
