+++ server/src/main/java/board/api/infra/mail/EmailMessage.java
@@ -1,0 +1,13 @@
+package board.api.infra.mail;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Builder
+@Data
+public class EmailMessage {
+
+    private String to;
+    private String subject;
+    private String text;
+}
