+++ server/src/main/java/board/api/infra/mail/EmailService.java
@@ -1,0 +1,8 @@
+package board.api.infra.mail;
+
+import javax.mail.MessagingException;
+
+public interface EmailService {
+
+    void send(EmailMessage emailMessage) throws MessagingException;
+}
