+++ server/src/main/java/board/api/modules/account/AccountRepository.java
@@ -1,0 +1,8 @@
+package board.api.modules.account;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AccountRepository extends JpaRepository<Account, Long> {
+
+    boolean existsByEmail(String email);
+}
