+++ server/src/main/java/board/api/modules/account/Role.java
@@ -1,0 +1,14 @@
+package board.api.modules.account;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Role {
+    USER("ROLE_USER", "회원"),
+    ADMIN("ROLE_ADMIN", "관리자");
+
+    private final String key;
+    private final String name;
+}
