+++ server/src/main/java/board/api/modules/articles/model/ArticleComment.java
@@ -1,0 +1,32 @@
+package board.api.modules.articles.model;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Getter
+@Setter
+@Table(name = "article_comment")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class ArticleComment extends BaseEntity {
+
+    @Id
+    @GeneratedValue
+    @Column(name = "article_comment_id")
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "article_id")
+    private Article article;
+
+    private String content;
+}
