+++ server/src/main/java/board/api/modules/articles/model/BaseEntity.java
@@ -1,0 +1,19 @@
+package board.api.modules.articles.model;
+
+import java.time.LocalDateTime;
+import javax.persistence.MappedSuperclass;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@MappedSuperclass
+public abstract class BaseEntity {
+
+    private LocalDateTime createdAt;
+
+    private String createdBy;
+
+    private LocalDateTime modifiedAt;
+    private String modifiedBy;
+}
