+++ server/src/main/java/board/api/modules/articles/model/Role.java
@@ -1,0 +1,5 @@
+package board.api.modules.articles.model;
+
+public enum Role {
+    ADMIN, USER
+}
