+++ server/src/main/java/board/api/modules/articles/service/ArticleService.java
@@ -1,0 +1,8 @@
+package board.api.modules.articles.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class ArticleService {
+
+}
