--- server/src/main/java/com/board/server/ServerApplication.java
@@ -1,17 +1,0 @@
-package com.board.server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@EnableJpaAuditing
-@ConfigurationPropertiesScan
-@SpringBootApplication
-public class ServerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ServerApplication.class, args);
-    }
-
-}
