--- server/src/main/java/com/board/server/config/AppConfig.java
@@ -1,21 +1,0 @@
-package com.board.server.config;
-
-import com.board.server.modules.member.mapper.AccountMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class AppConfig {
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
-    }
-
-    @Bean
-    public AccountMapper memberMapper() {
-        return new AccountMapper();
-    }
-}
