--- server/src/main/java/com/board/server/config/AppProperties.java
@@ -1,15 +1,0 @@
-package com.board.server.config;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.ConstructorBinding;
-
-@Getter
-@RequiredArgsConstructor
-@ConstructorBinding
-@ConfigurationProperties("app")
-public class AppProperties {
-
-    private final String host;
-}
