--- server/src/main/java/com/board/server/config/SecurityConfig.java
@@ -1,26 +1,0 @@
-package com.board.server.config;
-
-import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.builders.WebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-
-@Configuration
-public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        http.csrf().disable();
-
-        http.authorizeRequests()
-                .mvcMatchers("/", "/members/sign-up", "/login", "/members/authentication-mail").permitAll()
-                .anyRequest().authenticated();
-    }
-
-    @Override
-    public void configure(WebSecurity web) {
-        web.ignoring()
-                .requestMatchers(PathRequest.toStaticResources().atCommonLocations());
-    }
-}
