--- server/src/main/java/com/board/server/exception/InvalidSignUpRequestException.java
@@ -1,8 +1,0 @@
-package com.board.server.exception;
-
-public class InvalidSignUpRequestException extends IllegalArgumentException {
-
-    public InvalidSignUpRequestException(String message) {
-        super(message);
-    }
-}
