--- server/src/main/java/com/board/server/infra/mail/EmailMessage.java
@@ -1,13 +1,0 @@
-package com.board.server.infra.mail;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Builder
-@Data
-public class EmailMessage {
-
-    private String to;
-    private String subject;
-    private String text;
-}
