--- server/src/main/java/com/board/server/infra/mail/EmailService.java
@@ -1,8 +1,0 @@
-package com.board.server.infra.mail;
-
-import javax.mail.MessagingException;
-
-public interface EmailService {
-
-    void send(EmailMessage emailMessage) throws MessagingException;
-}
