--- server/src/main/java/com/board/server/modules/member/AccountRepository.java
@@ -1,8 +1,0 @@
-package com.board.server.modules.member;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AccountRepository extends JpaRepository<Account, Long> {
-
-    boolean existsByEmail(String email);
-}
