--- server/src/main/java/com/board/server/modules/member/Role.java
@@ -1,14 +1,0 @@
-package com.board.server.modules.member;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-    USER("ROLE_USER", "회원"),
-    ADMIN("ROLE_ADMIN", "관리자");
-
-    private final String key;
-    private final String name;
-}
