--- server/src/main/java/com/board/server/modules/member/mapper/AccountMapper.java
@@ -1,11 +1,0 @@
-package com.board.server.modules.member.mapper;
-
-import com.board.server.modules.member.Account;
-import com.board.server.modules.member.dto.SignUpResponseDto;
-
-public class AccountMapper {
-
-    public SignUpResponseDto toResponseDto(Account account) {
-        return new SignUpResponseDto(account);
-    }
-}
