+++ server/src/main/resources/application.properties
@@ -1,12 +1,14 @@
 spring.jpa.hibernate.ddl-auto=create
-spring.datasource.driver-class-name=
 spring.datasource.url=jdbc:mysql://localhost:3306/board
 spring.datasource.username=root
 spring.jpa.properties.hibernate.dialect-org.hibernate.dialect.MySQL5InnoDBDialect
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql = true
 
-spring.profiles.active=test
+logging.level.org.hibernate.SQL=debug
+logging.level.org.hibernate.type=trace
+
+spring.profiles.active= test
 app.host=http://localhost:8080
 
 management.endpoints.web.exposure.include=*
