+++ server/src/main/resources/templates/mail/authentication.html
@@ -1,12 +1,12 @@
 <!DOCTYPE html>
 <html lang="ko" xmlns:th="http://www.thymeleaf.org">
-  <head>
-    <meta charset="UTF-8">
-    <title>Elite King Board 회원가입 인증 메일</title>
-  </head>
+<head>
+  <meta charset="UTF-8">
+  <title>Elite King Board 회원가입 인증 메일</title>
+</head>
 
-  <body>
-    <p><span th:text="${nickname}">unknown</span>님 안녕하세요. 회원가입 인증 메일 입니다.</p>
-    <p>회원가입을 완료하기 위해서는 <a th:href="${link}">여기를 클릭해주세요.</a></p>
-  </body>
+<body>
+<p><span th:text="${nickname}">unknown</span>님 안녕하세요. 회원가입 인증 메일 입니다.</p>
+<p>회원가입을 완료하기 위해서는 <a th:href="${link}">여기를 클릭해주세요.</a></p>
+</body>
 </html>
